Updated the way actions work and commented
Before actions were called on the psuedo response object.  Now
they get passed both a psuedo response and request object, and they
are called on an 'Action' object.

The advantage of this approach is that the action object can now
keep track of the state of the object, and handle asyncronous calls
a lot easier.