Added contentType and encoding variables to lib/response.js and wrote tests for the whole module.
These were taken (with variations) from Steffen Tiedemann Christensen
http://github.com/steffentchr/bomberjs/commit/33d135ef37779dd569a1089dfd3f0e20a725272e
http://github.com/steffentchr/bomberjs/commit/66a4c79bcdb1361d513aa33902448e2c9bdf5565

Now, you can set the contentType and encoding on the response wrapper:

response.contentType = 'text/json';
response.encoding = 'utf8';

response.contentType will be used to set the "Content-Type" header if one
doesn't exist. It defaults to 'text/html'.

response.encoding will be used to specify the encoding used for the Node
serverResponse.sendBody method. It defualts to 'utf8'.

response.encoding will also be used to specify an encoding in the
"Content-Type" header if response.contentType is being used and it
starts with "text/".  If it starts otherwise I don't think we can make
that assumption.

While I was in there I also prefixed variables that are more internal
with an underscore.