Wrote tests for new Action implementation and fixed a few related bugs/inconsistencies
Before actions had this idea of "tasks".  Well that has been completely
scrapped and now we use promises for async code.

So, before an "action" might have looked like this:

exports.myAction = function(request, response) {
  this.addTask(function(request, response) {
      // do something here
    });
  this.addTask(function(request, response) {
      // do something else here and return a promise
      return promise;
    });
  this.addTask(function(request, response, resultOfPromise) {
      // do something with the result of the promise here
    });
}

Now it could look like the following (though it very much depends on
what asynchronous code you are working with, what you'd want
to do here).

exports.myAction = function(request, response) {
  var promise = new Promise();
  promise.addCallback(function() {
      // do something here
    });
  promise.addCallback(function(result1) {
      // do something else here and return a promise
      var promise2 = new Promise();
      promise2.addCallback(function(result2) {
          // do something with the result of the second promise here
        })
      return promise2;
    });

  return promise;
}

This will be most useful if using a library that utilizes Promises:

exports.myAction = function(request, response) {
  return db.query("SELECT STUFF").addCallback(function(db_response) {
      /* do something with database response and create page */
      return page;
    });
}

But it will still work with libraries that take callbacks:

exports.myAction = function(request, response) {
  var promise = new Promise();
  db.query("SELECT STUFF", function(db_response) {
      /* do something with database response and create page */
      promise.callback(page);
    });

  return promise;
}

How is this an improvement?  Well, we are now using what will be a
fundamental building block of Node for doing asynchronous code. So, it
should be a lot easier to get what should happen given you are used to
writing with Node.  Otherwise you have to learn a new system either way.

The previous Action implementation with its tasks was nothing but a
convoluted promise anyway, so why should we worry about duplicating and
maintaining that code?