Create Response.sendfile for sending the contents of a file back as a response.
This will read the file in chunks and send them back as they come in, so
it doesn't have to load the whole file into memory at once if it is a
large file.

Uses utils.mime to set the contentType if one isn't specified.

Heavily inspired by Steffen Tiedemann Christensen's serveFile.