Lots of changes to Response
Fixed sendFile.  Made it so it actually served the file properly.

Added a new charset property and stopped incorrectly using the encoding
property for this.  encoding is now 'transferEncoding' and is just used
to tell Node how to send data to the client.

Added tests for all of the above.