Change the way the cookie headers are set.
Instead of sending one 'Set-Cookie' header with a comma separated list
of cookies to set (which it looks like webkit can read fine, but not
Firefox), send the 'Set-Cookie' header multiple times, once for each
cookie.

This requires adding a little hack to Response. But Isaacs is working on
a fix for node so we don't have to do something like that.