Create a new object called "project" that will handle the global state of a set of Bomber app.
I decided that it was crucial that we have a place to store global
variables that any app can access and manipulate.  So, a database app
could add connections, or whatever.  Before I was thinking that the
server would satisfy this purpose but I didn't like the semantics of that.
The server should just be for serving requests.

So, we now have a global "project" object that all apps and the server
get a reference to.  Out of the box the project has a reference to

+ the base app that is being run
+ the server (if we are running the 'start-server' task)
+ the project config

As you can see I have introduced a new concept to the configuration, the
project config.  So, now we have two types of config, app_config and
project_config.  The difference is simple, app_config is for settings
that relate only to a specific instance of a specific app.  The project
config is for any other settings, or settings that the whole app can
access.

Here is an example config.js file, demonstrating how to set each.

  exports.project_config = {
    server: {
      port: 5456,
    },
    secret: 'this-is-my-secret',
  };

  exports.app_config = {
    '.': {
      option1: 1,
      option2: 2
    },
  };

  exports.apps = [
    'admin',
    'authentication'
  ];

Exactly where to put what settings is a little unclear at this point.

For example, a sessions app will modify the request objects globally in
the entire "project", so it seems like the settings for the session app
should belong in the project config, because you can't have app by app
differences in sessions.

We'll have to see if having this distinction is useful in the long run,
or if it only complicates things too much.  To truely make apps
plug-and-playable and nestable I think we need it though.